refactor(worker): extract handler into internal/worker package (#280)
Pure refactor with no behavior change in default subprocess build. Moves
workerImpl + helpers (stream bridging, drain, parse, options, error
classification, round-robin shared-state plumbing) out of cmd/worker/main.go
into internal/worker as a Handler type satisfying workerplugin.Worker.

cmd/worker/main.go becomes thin go-plugin glue: init runtime, load
defaults, build handler, serve. Process-global state (workerClientDefaults,
sharedStateClient atomic pointer, workerLogger, warning sync.Once values)
moves to Handler fields, making the handler constructible without hidden
package-global mutation.

Introduces SharedStateHook interface so the handler invokes shared-state
through an abstraction. Subprocess build wraps existing pb.SharedStateClient
via grpcSharedStateHook, preserving identical wire behavior. Also lands
the FetchAddStore + NewStoreSharedStateHook direct-call adapter, unused
in this PR but locking the seam consumed by PR 2.

IPC wire layer (workerplugin/*) is unchanged: same handshake, same
protobuf, same workerplugin.Worker interface, same RemoteAdvancer.
cmd/serve/* and pool/* are untouched. The //go:embed worker continues
to work identically.

Refs #280.
