+++ embed.go
@@ -15,7 +15,7 @@ import (
 	"github.com/invakid404/baml-rest/workerplugin"
 )
 
-//go:embed adapter.go adapters cmd/build cmd/embed cmd/hacks cmd/introspect/main.go cmd/schema/main.go cmd/serve/debug.go cmd/serve/debug_stub.go cmd/serve/error.go cmd/serve/headers.go cmd/serve/main.go cmd/serve/openapi.json cmd/serve/streamwriter.go cmd/serve/unary.go cmd/serve/unary_handlers.go cmd/serve/unary_stub.go cmd/serve/worker cmd/verify-adapter-pins cmd/verify-framework-adapter/main.go cmd/worker/error_classify.go cmd/worker/main.go embed.go go.mod go.sum go.work go.work.sum internal/apierror/error.go internal/httplogger internal/memlimit/memlimit.go internal/unsafeutil renovate.json scripts
+//go:embed adapter.go adapters cmd/build cmd/embed cmd/hacks cmd/introspect/main.go cmd/schema/main.go cmd/serve/debug.go cmd/serve/debug_stub.go cmd/serve/error.go cmd/serve/headers.go cmd/serve/main.go cmd/serve/openapi.json cmd/serve/streamwriter.go cmd/serve/unary.go cmd/serve/unary_handlers.go cmd/serve/unary_stub.go cmd/serve/worker cmd/verify-adapter-pins cmd/verify-framework-adapter/main.go cmd/worker embed.go go.mod go.sum go.work go.work.sum internal/apierror/error.go internal/httplogger internal/memlimit/memlimit.go internal/unsafeutil internal/worker/admin.go internal/worker/defaults.go internal/worker/errors.go internal/worker/handler.go internal/worker/metrics.go internal/worker/options.go internal/worker/parse.go internal/worker/runtime.go internal/worker/sharedstate.go internal/worker/stream.go renovate.json scripts
 var source embed.FS
 
 var Sources = make(map[string]embed.FS)
