+++ internal/worker/errors.go
@@ -1,4 +1,4 @@
-package main
+package worker
 
 import (
 	"errors"
