+++ internal/worker/errors_test.go
@@ -1,4 +1,4 @@
-package main
+package worker
 
 import (
 	"encoding/json"
